% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsedata.R
\name{getAugmentedParseData}
\alias{getAugmentedParseData}
\title{Key internal function fo sourcemodify
Returns parsed code as a tibble with detailed information
where each subexpression is found in the code}
\usage{
getAugmentedParseData(
  code,
  calls = parse(text = code, keep.source = TRUE),
  line.starts = find.line.starts(code)
)
}
\description{
Key internal function fo sourcemodify
Returns parsed code as a tibble with detailed information
where each subexpression is found in the code
}
