% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{somo_calls_info}
\alias{somo_calls_info}
\title{A simplified interface to somo_find_fun_arg}
\usage{
somo_calls_info(
  somo,
  fun,
  arg_names,
  arg_pos = NULL,
  arg_def = NULL,
  call_pd_rows = NULL,
  arg_prefix = ""
)
}
\description{
Only works with named arguments.
}
