% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_change_assign_rhs}
\alias{somo_change_assign_rhs}
\title{Changes the rhs of variable assignments}
\usage{
somo_change_assign_rhs(
  somo,
  id,
  new,
  new_df = data.frame(parent = id, new = new)
)
}
\arguments{
\item{somo}{The somo object generated with somo_init.}

\item{id}{The ids of the assignment as specified in somo$as_df}

\item{new}{New texts for the right hand sides}
}
\value{
The modified somo object
}
\description{
Take a look at somo$as_df to find the ids of the assignment
that shall be changed.
To get the new code call afterward somo_make_code
}
