% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_change_calls_fun_name}
\alias{somo_change_calls_fun_name}
\title{Change the name of a called function without adapting any arguments}
\usage{
somo_change_calls_fun_name(somo, fun, new, ignore_if_pkg = TRUE)
}
\arguments{
\item{somo}{The somo object generated with somo_init.}

\item{fun}{Function name that shall be changed in calls}

\item{new}{New function name to be called}

\item{ignore_if_pkg}{If TRUE don't change calls of the form pkg::fun(). If FALSE a call like pkg::fun() will be replaced by just new().}
}
\value{
The modified somo object
}
\description{
If the function call directly
}
