% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{somo_find_fun_arg}
\alias{somo_find_fun_arg}
\title{Find the text of the function argument in all calls to fun}
\usage{
somo_find_fun_arg(
  somo,
  fun,
  arg_name = NULL,
  arg_pos = NULL,
  arg_def = NULL,
  call_pd_rows = NULL
)
}
\description{
Find the text of the function argument in all calls to fun
}
