% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{somo_find_fun_call}
\alias{somo_find_fun_call}
\title{Find the text of the function argument in all calls to fun}
\usage{
somo_find_fun_call(somo, fun)
}
\description{
Find the text of the function argument in all calls to fun
}
