% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_make_mod_df}
\alias{somo_make_mod_df}
\title{Generate a data frame that contains information for all code modifications}
\usage{
somo_make_mod_df(somo)
}
\description{
Generate a data frame that contains information for all code modifications
}
