% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_make_pd_code}
\alias{somo_make_pd_code}
\title{Computes for the specified rows in somo$pd the modified code
after replacements have been performed.}
\usage{
somo_make_pd_code(somo, pd_rows = seq_len(NROW(somo$pd)))
}
\arguments{
\item{somo}{A somo object}

\item{pd_rows}{An integer vector of rows of somo$pd for which the modified code shall be computed.}
}
\value{
A character vector with resulting code for each row in pd specified by pd_rows
}
\description{
WARNING: Works currently only for non-nested replacements
}
