% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_surround_calls}
\alias{somo_surround_calls}
\title{Surround calls to the specified functions}
\usage{
somo_surround_calls(somo, funs = NULL, pre, post, call_pd_rows = NULL)
}
\arguments{
\item{somo}{The somo object generated with somo_init.}

\item{funs}{One or several function names.}

\item{pre}{Text to be added before the function call.}

\item{post}{Text to be added the function call.}
}
\value{
The modified somo object
}
\description{
Will by typically used to wrap a function call into
another function call that can analyse or modify the result.
}
