% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_surround_calls_arg}
\alias{somo_surround_calls_arg}
\title{Surround an argument of specific function}
\usage{
somo_surround_calls_arg(
  somo,
  fun = NULL,
  arg_name = NULL,
  arg_pos = NULL,
  pre,
  post,
  arg_def = NULL,
  call_pd_rows = NULL
)
}
\arguments{
\item{somo}{The somo object generated with somo_init.}

\item{fun}{Name of the function whose argument shall be wrapped}

\item{arg_name}{The name of the argument (must be given if arg_pos is NULL)}

\item{arg_pos}{Position of the argument (must be given if arg_name is NULL)}

\item{pre}{Text to be added before the function call.}

\item{post}{Text to be added the function call.}

\item{arg_def}{Names of all function arguments in the correct order. The result of calling names(as.list(args(fun))). This argument is optional, but without it matching may fail.}
}
\value{
The modified somo object
}
\description{
Will by typically used to wrap the argument (like a filenam)
of function call into another function call
that can analyse or modify the argument.
}
