% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcemodify.R
\name{somo_init}
\alias{somo_init}
\title{Init a new source modifier object}
\usage{
somo_init(code, start.size = 200)
}
\arguments{
\item{code}{The R source code as text file}
}
\description{
Init a new source modifier object
}
